change: moved get_file_extension function to core/common/utils, updated imports
