+++ core/common/utils.py
@@ -1,0 +1,5 @@
+def get_file_extension(filename):
+    """
+    Get extension of a file in .ext format
+    """
+    return '.' + filename.split('.')[-1]
